workflow: Fix indentation of event triggers in the system tests workflow
The current configuration file has incorrect indentation for 'release' and
'workflow_dispatch' events. They are in the scope of 'pull_request' due to
the indentation level. This leads to misinterpretation of the configuration
file, as these events are expected to be top-level, similar to 'push' and
'pull_request'.

This commit fix the formatting, now 'release' and 'workflow_dispatch' are
independent event triggers.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>