workflow: Fix checkout action for PRs from forks
Currently, the checkout action in the system tests workflow doesn't handle all
possible GitHub events correctly. For example, when a pull request is created
from a fork, the action fails because it can't find the
branch in the base repository.

This commit enhances the Checkout action to handle different GitHub
events such as 'push', 'pull_request' and 'release'. It uses the
appropriate repository and reference based on the event that triggered
the workflow.

For pull requests, the action now checks out the correct repository and branch,
even when the PR comes from a fork. For push events, it checks out the branch
that was pushed. For releases, it checks out the tag that was released.

Additionally, the 'workflow_dispatch' event is also handled, allowing
users to manually trigger the workflow and specify a reference.

Signed-off-by: Oleg Kulachenko <oleg@nspcc.ru>